requirements.txt
python-dotenv==1.0.1
pytest==8.2.2
pytest-cov==5.0.0
httpx==0.27.0
requests==2.31.0
openai==1.3.0
python-dateutil==2.8.2
Flask==2.2.2
Werkzeug==2.2.3
imaplib2==3.6
email-validator==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai~=0.5.0
black==23.11.0
isort==5.13.2
